Move several functions on AggregateType from scopeResolve.{h,cpp} to type.{h,cpp}
