+++ compiler/AST/type.cpp
@@ -33,6 +33,7 @@
 #include "iterator.h"
 #include "misc.h"
 #include "passes.h"
+#include "scopeResolve.h"
 #include "stlUtil.h"
 #include "stringutil.h"
 #include "symbol.h"
@@ -1078,6 +1079,845 @@ std::string AggregateType::docsDirective() {
   return "";
 }
 
+
+/************************************* | **************************************
+*                                                                             *
+*                                                                             *
+************************************** | *************************************/
+
+static void       build_type_constructor(AggregateType* ct);
+static void       build_constructor(AggregateType* ct);
+static ArgSymbol* create_generic_arg(VarSymbol* field);
+static void       insert_implicit_this(FnSymbol*         fn,
+                                       Vec<const char*>& fieldNamesSet);
+static void       move_constructor_to_outer(FnSymbol*      fn,
+                                            AggregateType* outerType);
+
+void build_constructors(AggregateType* ct)
+{
+  if (ct->defaultInitializer)
+    return;
+
+  SET_LINENO(ct);
+
+  build_type_constructor(ct);
+  build_constructor(ct);
+}
+
+
+// Create the (default) type constructor for this class.
+static void build_type_constructor(AggregateType* ct) {
+  // Do nothing if it is already built
+  if (ct->defaultTypeConstructor)
+    return;
+
+  // Create the type constructor function,
+  FnSymbol* fn = new FnSymbol(astr("_type_construct_", ct->symbol->name));
+
+  fn->addFlag(FLAG_TYPE_CONSTRUCTOR);
+  fn->cname = astr("_type_construct_", ct->symbol->cname);
+
+  fn->addFlag(FLAG_COMPILER_GENERATED);
+  fn->retTag = RET_TYPE;
+
+  if (ct->symbol->hasFlag(FLAG_REF))
+    fn->addFlag(FLAG_REF);
+
+  if (ct->symbol->hasFlag(FLAG_TUPLE)) {
+    fn->addFlag(FLAG_TUPLE);
+    fn->addFlag(FLAG_INLINE);
+    gGenericTupleTypeCtor = fn;
+  }
+
+  // and insert it into the class type.
+  ct->defaultTypeConstructor = fn;
+
+  // Create "this".
+  fn->_this = new VarSymbol("this", ct);
+  fn->_this->addFlag(FLAG_ARG_THIS);
+
+  fn->insertAtTail(new DefExpr(fn->_this));
+
+  Vec<const char*> fieldNamesSet;
+
+  CallExpr* superCall = NULL;
+
+  // Copy arguments from superclass type constructor
+  // (supporting inheritance from generic classes)
+  if (isClass(ct) && ct->dispatchParents.n > 0) {
+
+    if(AggregateType *parentTy = toAggregateType(ct->dispatchParents.v[0])){
+
+      // This class/record has a parent class/record
+      if (!parentTy->defaultTypeConstructor) {
+        // If it doesn't yet have an type constructor, make one
+        build_type_constructor(parentTy);
+      }
+      FnSymbol* superTypeCtor = parentTy->defaultTypeConstructor;
+
+      if (superTypeCtor->numFormals() > 0) {
+
+        superCall = new CallExpr(parentTy->symbol->name);
+
+        // Now walk through arguments in super class type constructor
+        for_formals(formal, superTypeCtor) {
+
+          DefExpr* superArg = formal->defPoint->copy();
+
+          // Add a formal to the current class type constructor.
+
+          ArgSymbol* arg = toArgSymbol(superArg->sym->copy());
+          bool fieldInThisClass = false;
+          for_fields(sym, ct) {
+            if (0 == strcmp(sym->name, arg->name)) {
+              fieldInThisClass = true;
+            }
+          }
+
+          if (fieldInThisClass) {
+            // If the field is also present in the child, adjust the field
+            // name in the super. Otherwise it would not be possible to
+            // type construct the super.
+
+            // ?
+            // Should we omit them?
+            // Or pass the child field to the super constructor?
+            continue;
+          }
+          arg->addFlag(FLAG_PARENT_FIELD);
+          fn->insertFormalAtTail(arg);
+          superCall->insertAtTail(new SymExpr(arg));
+        }
+      }
+    }
+  }
+
+  for_fields(tmp, ct) {
+    SET_LINENO(tmp);
+
+    if (VarSymbol* field = toVarSymbol(tmp)) {
+      if (field->hasFlag(FLAG_SUPER_CLASS)) {
+        // supporting inheritance from generic classes
+        if (superCall) {
+          CallExpr* newInit = new CallExpr(PRIM_TYPE_INIT, superCall);
+          CallExpr* newSet  = new CallExpr(PRIM_SET_MEMBER,
+                                           fn->_this,
+                                           new_CStringSymbol(field->name),
+                                           newInit);
+          fn->insertAtTail(newSet);
+        }
+        continue;
+      }
+
+
+      Expr* exprType = field->defPoint->exprType;
+      Expr* init = field->defPoint->init;
+
+      if (!strcmp(field->name, "_promotionType")) {
+
+        fn->insertAtTail(
+          new BlockStmt(
+            new CallExpr(PRIM_SET_MEMBER, fn->_this, 
+              new_CStringSymbol(field->name),
+              new CallExpr(PRIM_TYPE_INIT, exprType->remove())),
+            BLOCK_TYPE));
+
+      } else {
+        fieldNamesSet.set_add(field->name);
+
+        //
+        // if formal is generic
+        //
+        if (field->isType() || 
+            field->hasFlag(FLAG_PARAM)         || 
+            (!exprType && !init)) {
+
+          ArgSymbol* arg = create_generic_arg(field);
+
+          // Indicate which type constructor args are also for super class
+          // This helps us to call the superclass type constructor in resolution
+          if (field->hasFlag(FLAG_PARENT_FIELD)) {
+            arg->addFlag(FLAG_PARENT_FIELD);
+          }
+
+          fn->insertFormalAtTail(arg);
+
+          if (field->hasFlag(FLAG_PARAM) || field->isType())
+            fn->insertAtTail(new CallExpr(PRIM_SET_MEMBER,
+                                          fn->_this,
+                                          new_CStringSymbol(field->name), arg));
+
+          else if (arg->type == dtAny &&
+                   !ct->symbol->hasFlag(FLAG_REF))
+            // It would be nice to be able to remove this case.
+            fn->insertAtTail(new CallExpr(PRIM_SET_MEMBER,
+                                          fn->_this,
+                                          new_CStringSymbol(field->name),
+                                          new CallExpr(PRIM_TYPE_INIT, arg)));
+        } else if (exprType) {
+          CallExpr* newInit = new CallExpr(PRIM_TYPE_INIT, exprType->copy());
+          CallExpr* newSet  = new CallExpr(PRIM_SET_MEMBER, 
+                                           fn->_this,
+                                           new_CStringSymbol(field->name),
+                                           newInit);
+          fn->insertAtTail(newSet);
+
+        } else if (init) {
+          fn->insertAtTail(new CallExpr(PRIM_SET_MEMBER,
+                                        fn->_this,
+                                        new_CStringSymbol(field->name),
+                                        new CallExpr("chpl__initCopy", init->copy())));
+        }
+      }
+    }
+  }
+
+  // Add return
+  fn->insertAtTail(new CallExpr(PRIM_RETURN, fn->_this));
+  fn->retType = ct;
+
+  ct->symbol->defPoint->insertBefore(new DefExpr(fn));
+
+  // Make implicit references to 'this' explicit.
+  insert_implicit_this(fn, fieldNamesSet);
+
+  AggregateType *outerType = toAggregateType(ct->symbol->defPoint->parentSymbol->type);
+
+  if (outerType) {
+    // Create an "outer" pointer to the outer class in the inner class
+    VarSymbol* outer = new VarSymbol("outer");
+
+    move_constructor_to_outer(fn, outerType);
+
+    // Save the pointer to the outer class
+    ct->fields.insertAtTail(new DefExpr(outer));
+    fn->insertAtHead(new CallExpr(PRIM_SET_MEMBER,
+                                  fn->_this,
+                                  new_CStringSymbol("outer"),
+                                  fn->_outer));
+
+    ct->outer = outer;
+  }
+
+  // Update the symbol table with added defs.
+  std::vector<DefExpr*> defs;
+
+  collectDefExprs(fn, defs);
+  addToSymbolTable(defs);
+}
+
+
+// For the given class type, this builds the compiler-generated constructor
+// which is also called by user-defined constructors to pre-initialize all
+// fields to their declared or type-specific initial values.
+static void build_constructor(AggregateType* ct) {
+  if (ct->initializerStyle == DEFINES_INITIALIZER) {
+    // Don't want to create the default constructor if we have seen initializers
+    // defined.  The work is completely unnecessary, since we won't call the
+    // default constructor, and it mutates information about the fields that
+    // we would rather stayed unmutated.
+    return;
+  }
+
+  // Create the default constructor function symbol,
+  FnSymbol* fn = new FnSymbol(astr("_construct_", ct->symbol->name));
+
+  fn->cname = fn->name;
+
+  fn->addFlag(FLAG_DEFAULT_CONSTRUCTOR);
+  fn->addFlag(FLAG_CONSTRUCTOR);
+  fn->addFlag(FLAG_COMPILER_GENERATED);
+
+  if (ct->symbol->hasFlag(FLAG_REF))
+    fn->addFlag(FLAG_REF);
+
+  if (ct->symbol->hasFlag(FLAG_TUPLE)) {
+    fn->addFlag(FLAG_TUPLE);
+    fn->addFlag(FLAG_INLINE);
+    gGenericTupleInit = fn;
+  }
+
+  // And insert it into the class type.
+  ct->defaultInitializer = fn;
+
+  // Create "this".
+  fn->_this = new VarSymbol("this", ct);
+  fn->_this->addFlag(FLAG_ARG_THIS);
+
+  fn->insertAtTail(new DefExpr(fn->_this));
+
+  // Walk the fields in the class type.
+  std::map<VarSymbol*,ArgSymbol*> fieldArgMap;
+
+  Vec<const char*> fieldNamesSet;
+
+  for_fields(tmp, ct) {
+    SET_LINENO(tmp);
+    if (VarSymbol* field = toVarSymbol(tmp)) {
+      // Filter inherited fields and other special cases.
+      // "outer" is used internally to supply a pointer to
+      // the outer parent of a nested class.
+      if (!field->hasFlag(FLAG_SUPER_CLASS) &&
+          strcmp(field->name, "_promotionType") &&
+          strcmp(field->name, "outer")) {
+        // Create an argument to the default constructor
+        // corresponding to the field.
+        ArgSymbol* arg = new ArgSymbol(INTENT_BLANK, field->name, field->type);
+
+        fieldArgMap[field] = arg;
+        fieldNamesSet.set_add(field->name);
+      }
+    }
+  }
+
+  ArgSymbol* meme      = NULL;
+  CallExpr*  superCall = NULL;
+  CallExpr*  allocCall = NULL;
+
+  if (ct->symbol->hasFlag(FLAG_REF)) {
+    // For ref, sync and single classes, just allocate space.
+    allocCall = callChplHereAlloc(fn->_this->type);
+
+    fn->insertAtTail(new CallExpr(PRIM_MOVE, fn->_this, allocCall));
+
+  } else if (!ct->symbol->hasFlag(FLAG_TUPLE)) {
+    // Create a meme (whatever that is).
+    meme = new ArgSymbol(INTENT_BLANK,
+                         "meme",
+                         ct,
+                         NULL,
+                         new SymExpr(gTypeDefaultToken));
+
+    meme->addFlag(FLAG_IS_MEME);
+
+    // Move the meme into "this".
+    fn->insertAtTail(new CallExpr(PRIM_MOVE, fn->_this, meme));
+
+    if (isClass(ct)) {
+      if (ct->dispatchParents.n > 0 && !ct->symbol->hasFlag(FLAG_EXTERN)) {
+        // This class has a parent class.
+        if (!ct->dispatchParents.v[0]->defaultInitializer) {
+          // If it doesn't yet have an initializer, make one.
+          build_constructors(toAggregateType(ct->dispatchParents.v[0]));
+        }
+
+        // Get the parent constructor.
+        // Note that since we only pay attention to the first entry in the
+        // dispatchParents list, we are effectively implementing
+        // single class inheritance, multiple interface inheritance.
+        FnSymbol* superCtor = ct->dispatchParents.v[0]->defaultInitializer;
+
+        // Create a call to the superclass constructor.
+        superCall = new CallExpr(superCtor->name);
+
+        // Walk the formals of the default super class constructor
+        for_formals_backward(formal, superCtor) {
+          if (formal->hasFlag(FLAG_IS_MEME))
+            continue;
+
+          DefExpr* superArg = formal->defPoint->copy();
+
+          // Omit the arguments shadowed by this class's fields.
+          if (fieldNamesSet.set_in(superArg->sym->name))
+            continue;
+
+          fieldNamesSet.set_add(superArg->sym->name);
+
+          // Inserting each successive ancestor argument at the head in
+          // reverse-lexical order results in all of the arguments appearing
+          // in lexical order, starting with those in the most ancient class
+          // and ending with those in the most-derived class.
+          fn->insertFormalAtHead(superArg);
+          superCall->insertAtHead(superArg->sym);
+        }
+
+        // Create a temp variable and add it to the actual argument list
+        // in the superclass constructor call.  This temp will hold
+        // the pointer to the parent subobject.
+        VarSymbol* tmp = newTemp();
+
+        superCall->insertAtTail(new NamedExpr("meme", new SymExpr(tmp)));
+
+        // Add super call to the constructor function.
+        fn->insertAtTail(superCall);
+
+        // Declare that variable in the scope of this constructor.
+        // And initialize it with the super class pointer.
+        superCall->insertBefore(new DefExpr(tmp));
+
+        superCall->insertBefore(
+          new CallExpr(PRIM_MOVE,
+                       tmp,
+                       new CallExpr(PRIM_GET_MEMBER_VALUE,
+                                    fn->_this,
+                                    new_CStringSymbol("super"))));
+      }
+    }
+  }
+
+  if (isUnion(ct))
+    fn->insertAtTail(new CallExpr(PRIM_SET_UNION_ID, fn->_this, new_IntSymbol(0)));
+
+  ct->symbol->defPoint->insertBefore(new DefExpr(fn));
+
+  for_fields(tmp, ct) {
+    VarSymbol* field = toVarSymbol(tmp);
+
+    if (!field)
+      continue;
+
+    if (fieldArgMap.count(field) == 0)
+      continue;
+
+    ArgSymbol* arg = fieldArgMap[field];
+
+    SET_LINENO(field);
+
+    if (field->hasFlag(FLAG_PARAM))
+      arg->intent = INTENT_PARAM;
+
+    Expr* exprType = field->defPoint->exprType;
+    Expr* init     = field->defPoint->init;
+
+    bool  hadType  = exprType;
+    bool  hadInit  = init;
+
+    if (exprType != NULL) {
+      exprType->remove();
+    }
+
+    if (init != NULL) {
+      init->remove();
+    }
+
+    if (init) {
+      if (!field->isType() && !exprType) {
+        // init && !exprType
+        VarSymbol* tmp = newTemp();
+
+        tmp->addFlag(FLAG_INSERT_AUTO_DESTROY);
+        tmp->addFlag(FLAG_MAYBE_PARAM);
+        tmp->addFlag(FLAG_MAYBE_TYPE);
+
+        exprType = new BlockStmt(new DefExpr(tmp), BLOCK_TYPE);
+
+        toBlockStmt(exprType)->insertAtTail(new CallExpr(PRIM_MOVE,
+                                                         tmp,
+                                                         new CallExpr("chpl__initCopy",
+                                                                      init->copy())));
+
+        toBlockStmt(exprType)->insertAtTail(new CallExpr(PRIM_TYPEOF, tmp));
+      }
+    } else if (hadType &&
+               !field->isType() &&
+               !field->hasFlag(FLAG_PARAM)) {
+      init = new CallExpr(PRIM_INIT, exprType->copy());
+    }
+
+
+    if (!field->isType() && !field->hasFlag(FLAG_PARAM)) {
+      if (hadType) {
+        init = new CallExpr("_createFieldDefault", exprType->copy(), init);
+      } else if (init)
+        init = new CallExpr("chpl__initCopy", init);
+    }
+
+    if (exprType) {
+      if (!isBlockStmt(exprType))
+        arg->typeExpr = new BlockStmt(exprType, BLOCK_TYPE);
+      else
+        arg->typeExpr = toBlockStmt(exprType);
+    }
+
+    if (init) {
+      if (hadInit)
+        arg->defaultExpr = new BlockStmt(init, BLOCK_SCOPELESS);
+      else {
+        Expr* initVal = new SymExpr(gTypeDefaultToken);
+        arg->defaultExpr = new BlockStmt(initVal);
+      }
+    }
+
+    if (field->isType())
+      // Args with this flag are removed after resolution.
+      // Note that in the default type constructor, this flag is also applied
+      // (along with FLAG_GENERIC) to arguments whose type is unknown, but would
+      // not be pruned in resolution.
+      arg->addFlag(FLAG_TYPE_VARIABLE);
+
+    if (!exprType && arg->type == dtUnknown)
+      arg->type = dtAny;
+
+    fn->insertFormalAtTail(arg);
+
+    if (arg->type == dtAny && !arg->hasFlag(FLAG_TYPE_VARIABLE) &&
+        !arg->hasFlag(FLAG_PARAM) && !ct->symbol->hasFlag(FLAG_REF))
+      fn->insertAtTail(new CallExpr(PRIM_SET_MEMBER, 
+                                    fn->_this, 
+                                    new_CStringSymbol(arg->name),
+                                    new CallExpr("chpl__initCopy", arg)));
+    else
+      // Since we don't copy the argument before stuffing it in a field,
+      // we will have to remove the autodestroy flag for specific cases.
+      // Namely, if the function is a default constructor and the target
+      // of a PRIM_SET_MEMBER is a record, then the INSERT_AUTO_DESTROY
+      // flag must be removed.
+      // (See NOTE 1 in callDestructors.cpp.)
+      fn->insertAtTail(new CallExpr(PRIM_SET_MEMBER, 
+                                    fn->_this, 
+                                    new_CStringSymbol(arg->name),
+                                    arg));
+  }
+
+  if (meme)
+    fn->insertFormalAtTail(meme);
+
+  insert_implicit_this(fn, fieldNamesSet);
+
+  AggregateType *outerType = toAggregateType(ct->symbol->defPoint->parentSymbol->type);
+
+  if (outerType) {
+    move_constructor_to_outer(fn, outerType);
+
+    // Save the pointer to the outer class
+    fn->insertAtTail(new CallExpr(PRIM_SET_MEMBER,
+                                  fn->_this,
+                                  new_CStringSymbol("outer"),
+                                  fn->_outer));
+  }
+
+  //
+  // Insert a call to the "initialize()" method if one is defined.
+  // The return value of this method (if any) is ignored.
+  //
+  forv_Vec(FnSymbol, method, ct->methods) {
+    // Select a method named "initialize" and taking no arguments
+    // (aside from _mt and the implicit 'this').
+    if (method && !strcmp(method->name, "initialize")) {
+      if (method->numFormals() == 2) {
+        CallExpr* init = new CallExpr("initialize", gMethodToken, fn->_this);
+        fn->insertAtTail(init);
+        // If a record type has an initialize method, it's not Plain Old Data.
+        if (!isClass(ct)) {
+          ct->symbol->addFlag(FLAG_NOT_POD);
+        }
+        break;
+      }
+    }
+  }
+
+  fn->insertAtTail(new CallExpr(PRIM_RETURN, fn->_this));
+
+  std::vector<DefExpr*> defs;
+  collectDefExprs(fn, defs);
+  addToSymbolTable(defs);
+}
+
+static ArgSymbol* create_generic_arg(VarSymbol* field)
+{
+  ArgSymbol* arg = new ArgSymbol(INTENT_BLANK, field->name, field->type);
+
+  // We take it as a param argument if it is marked as a param field.
+  if (field->hasFlag(FLAG_PARAM))
+    arg->intent = INTENT_PARAM;
+  else
+    // Both type arguments and arguments of unspecified type get this flag.
+    arg->addFlag(FLAG_TYPE_VARIABLE);
+
+  // Copy the field type if it exists.
+  Expr* exprType = field->defPoint->exprType;
+  if (exprType)
+    arg->typeExpr = new BlockStmt(exprType->copy(), BLOCK_TYPE);
+
+  // Copy the initialization expression if it exists.
+  Expr* init = field->defPoint->init;
+  if (init)
+    arg->defaultExpr = new BlockStmt(init->copy(), BLOCK_SCOPELESS);
+
+  // Translate an unknown field type into an unspecified arg type.
+  if (!exprType && arg->type == dtUnknown)
+  {
+    if (! field->isType())
+      arg->addFlag(FLAG_GENERIC);
+    arg->type = dtAny;
+  }
+
+  return arg;
+}
+
+/// Replace implicit references to 'this' in the body of this 
+/// type constructor with explicit member reference (dot) expressions.
+static void insert_implicit_this(FnSymbol* fn, Vec<const char*>& fieldNamesSet)
+{
+  std::vector<BaseAST*> asts;
+
+  collect_asts(fn->body, asts);
+
+  for_vector(BaseAST, ast, asts) {
+    if (UnresolvedSymExpr* se = toUnresolvedSymExpr(ast))
+      if (fieldNamesSet.set_in(se->unresolved))
+        // The name of this UnresolvedSymExpr matches a field name.
+        // So replace it with a dot expression.
+        se->replace(buildDotExpr(fn->_this, se->unresolved));
+  }
+}
+
+static void move_constructor_to_outer(FnSymbol* fn, AggregateType* outerType)
+{
+  // Remove the DefPoint for this constructor, add it to the outer
+  // class's method list.
+  outerType->methods.add(fn);
+
+  fn->_outer = new ArgSymbol(INTENT_BLANK, "outer", outerType);
+  fn->_outer->addFlag(FLAG_GENERIC); // Arg expects a real object :-P.
+  fn->insertFormalAtHead(new DefExpr(fn->_outer));
+  fn->insertFormalAtHead(new DefExpr(new ArgSymbol(INTENT_BLANK,
+                                                   "_mt",
+                                                   dtMethodToken)));
+  fn->addFlag(FLAG_METHOD);
+
+  fn->addFlag(FLAG_METHOD_PRIMARY);
+
+  Expr* insertPoint = outerType->symbol->defPoint;
+
+  while (toTypeSymbol(insertPoint->parentSymbol))
+    insertPoint = insertPoint->parentSymbol->defPoint;
+
+  insertPoint->insertBefore(fn->defPoint->remove());
+}
+
+/************************************* | **************************************
+*                                                                             *
+* Compute dispatchParents and dispatchChildren vectors; add base class fields *
+* to subclasses; identify cyclic or illegal class or record hierarchies       *
+*                                                                             *
+************************************** | *************************************/
+
+static void           addClassToHierarchy(AggregateType*            ct,
+                                          std::set<AggregateType*>& seen);
+
+static AggregateType* discoverParentAndCheck(Expr*          storesName,
+                                             AggregateType* child);
+
+void addClassToHierarchy(AggregateType* ct) {
+  std::set<AggregateType*> localSeen; // classes in potential cycle
+
+  addClassToHierarchy(ct, localSeen);
+}
+
+static void addClassToHierarchy(AggregateType*            ct,
+                                std::set<AggregateType*>& localSeen) {
+  // classes already in hierarchy
+  static std::set<AggregateType*> globalSeen;
+
+  if (localSeen.find(ct)  != localSeen.end())  {
+    USR_FATAL(ct, "Class hierarchy is cyclic");
+  }
+
+  if (globalSeen.find(ct) != globalSeen.end()) {
+    return;
+  }
+
+  globalSeen.insert(ct);
+
+  add_root_type(ct);
+
+  // Walk the base class list, and add parents into the class hierarchy.
+  for_alist(expr, ct->inherits) {
+    AggregateType* pt = discoverParentAndCheck(expr, ct);
+
+    localSeen.insert(ct);
+
+    addClassToHierarchy(pt, localSeen);
+
+    ct->dispatchParents.add(pt);
+
+    if (ct->isGeneric() == false && pt->isGeneric() == true) {
+      ct->markAsGeneric();
+    }
+
+    if (pt->dispatchChildren.add_exclusive(ct) == false) {
+      INT_ASSERT(false);
+    }
+
+    expr->remove();
+
+    if (isClass(ct) == true) {
+      SET_LINENO(ct);
+
+      // For a class, just add a super class pointer.
+      VarSymbol* super = new VarSymbol("super", pt);
+
+      super->addFlag(FLAG_SUPER_CLASS);
+
+      ct->fields.insertAtHead(new DefExpr(super));
+
+    } else {
+      SET_LINENO(ct);
+
+      // For records and unions, scan the fields in the parent type.
+      for_fields_backward(field, pt) {
+        if (toVarSymbol(field) && !field->hasFlag(FLAG_SUPER_CLASS)) {
+          // If not already in derived class (by name), copy it.
+          bool alreadyContainsField = false;
+
+          for_fields(myfield, ct) {
+            if (strcmp(myfield->name, field->name) == 0) {
+              alreadyContainsField = true;
+              break;
+            }
+          }
+
+          if (!alreadyContainsField) {
+            DefExpr* def = field->defPoint->copy();
+
+            ct->fields.insertAtHead(def);
+
+            def->sym->addFlag(FLAG_PARENT_FIELD);
+          }
+        }
+      }
+    }
+  }
+}
+
+static AggregateType* discoverParentAndCheck(Expr*          storesName,
+                                             AggregateType* child) {
+  UnresolvedSymExpr* se  = toUnresolvedSymExpr(storesName);
+  Symbol*            sym = lookup(storesName, se->unresolved);
+  TypeSymbol*        ts  = toTypeSymbol(sym);
+
+  //    printf("looking up %s\n", se->unresolved);
+  if (!ts) {
+    USR_FATAL(storesName, "Illegal super class");
+  }
+
+  //    printf("found it in %s\n", sym->getModule()->name);
+  AggregateType* pt = toAggregateType(ts->type);
+
+  if (!pt) {
+    USR_FATAL(storesName, "Illegal super class %s", ts->name);
+  }
+
+  if (isUnion(child) && isUnion(pt)) {
+    USR_FATAL(storesName, "Illegal inheritance involving union type");
+  }
+
+  if (isRecord(child) && isClass(pt)) {
+    USR_FATAL(storesName,
+              "Record %s inherits from class %s",
+              child->symbol->name,
+              pt->symbol->name);
+  }
+
+  if (isClass(child) && isRecord(pt)) {
+    // <hilde> Possible language change: Allow classes to inherit
+    // fields and methods from records.
+    USR_FATAL(storesName,
+              "Class %s inherits from record %s",
+              child->symbol->name,
+              pt->symbol->name);
+  }
+
+  return pt;
+}
+
+void addRecordDefaultConstruction() {
+  forv_Vec(DefExpr, def, gDefExprs) {
+    // We're only interested in declarations that do not have initializers.
+    if (def->init != NULL) {
+
+    } else if (VarSymbol* var = toVarSymbol(def->sym)) {
+      if (AggregateType* at = toAggregateType(var->type)) {
+        if (at->isRecord() == false) {
+
+        // No initializer for extern records.
+        } else if (at->symbol->hasFlag(FLAG_EXTERN) == true) {
+
+        } else {
+          SET_LINENO(def);
+
+          CallExpr* ctor_call = new CallExpr(new SymExpr(at->symbol));
+
+          def->init = new CallExpr(PRIM_NEW, ctor_call);
+
+          insert_help(def->init, def, def->parentSymbol);
+        }
+      }
+    }
+  }
+}
+
+void setCreationStyle(TypeSymbol* t, FnSymbol* fn) {
+  bool isCtor = (0 == strcmp(t->name, fn->name));
+  bool isInit = (0 == strcmp(fn->name, "init"));
+
+  if (!isCtor && !isInit)
+    return;
+
+  AggregateType* ct = toAggregateType(t->type);
+  if (!ct)
+    INT_FATAL(fn, "initializer on non-class type");
+
+  if (fn->hasFlag(FLAG_NO_PARENS)) {
+    USR_FATAL(fn, "a%s cannot be declared without parentheses", isCtor ? " constructor" : "n initializer");
+  }
+
+  if (ct->initializerStyle == DEFINES_NONE_USE_DEFAULT) {
+    // We hadn't previously seen a constructor or initializer definition.
+    // Update the field on the type appropriately.
+    if (isInit) {
+      ct->initializerStyle = DEFINES_INITIALIZER;
+    } else if (isCtor) {
+      ct->initializerStyle = DEFINES_CONSTRUCTOR;
+    } else {
+      // Should never reach here, but just in case...
+      INT_FATAL(fn, "Function was neither a constructor nor an initializer");
+    }
+  } else if ((ct->initializerStyle == DEFINES_CONSTRUCTOR && !isCtor) ||
+             (ct->initializerStyle == DEFINES_INITIALIZER && !isInit)) {
+    // We've previously seen a constructor but this new method is an initializer
+    // or we've previously seen an initializer but this new method is a
+    // constructor.  We don't allow both to be defined on a type.
+    USR_FATAL_CONT(fn, "Definition of both constructor '%s' and initializer 'init'.  Please choose one.", ct->symbol->name);
+  }
+}
+
+/************************************* | **************************************
+*                                                                             *
+*                                                                             *
+************************************** | *************************************/
+
+void add_root_type(AggregateType* ct) {
+  // make root records inherit from value
+  // make root classes inherit from object
+  if (ct->inherits.length == 0 && !ct->symbol->hasFlag(FLAG_NO_OBJECT)) {
+    SET_LINENO(ct);
+
+    if (isRecord(ct)) {
+      ct->dispatchParents.add(dtValue);
+
+      // Assume that this addition is unique; report if not.
+      if (dtValue->dispatchChildren.add_exclusive(ct) == false) {
+        INT_ASSERT(false);
+      }
+
+    } else if (isClass(ct)) {
+      ct->dispatchParents.add(dtObject);
+
+      // Assume that this addition is unique; report if not.
+      if (dtObject->dispatchChildren.add_exclusive(ct) == false) {
+        INT_ASSERT(false);
+      }
+
+      VarSymbol* super = new VarSymbol("super", dtObject);
+
+      super->addFlag(FLAG_SUPER_CLASS);
+
+      ct->fields.insertAtHead(new DefExpr(super));
+    }
+  }
+}
 
 /************************************* | **************************************
 *                                                                             *
